requirements.txt
python-whois==0.7.0
requests==2.18.4